requirements.txt
PyQt6>=6.4.0
paho-mqtt>=2.0.0
